pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amazon-bearing-price-monitor"
version = "0.1.0"
description = "Real-time Amazon India bearing price monitoring system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "asyncpg>=0.29.0",
  "celery[redis]>=5.4.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "motor>=3.5.0",
  "orjson>=3.10.0",
  "playwright>=1.45.0",
  "prometheus-client>=0.20.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.3.0",
  "scrapy-playwright>=0.0.35",
  "Scrapy>=2.11.0",
  "sqlalchemy[asyncio]>=2.0.31",
  "sse-starlette>=2.1.2",
  "structlog>=24.2.0",
  "tenacity>=8.4.2",
  "uvicorn[standard]>=0.30.1",
  "python-slugify>=8.0.4",
]

[project.optional-dependencies]
dev = [
  "black>=24.4.2",
  "flake8>=7.1.0",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*", "scraper*"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests", "scraper/tests"]
